app/requirements.txt
Flask==2.1.2
matplotlib==3.3.3
Pillow==9.3.0
torch==1.10.2
torchvision==0.11.3
Werkzeug==2.1.2
